1.2
Added a new menu to the city with the following things: A option where you can restore your health.

A new shop system that includes: Pets, health potions and 4 different pieces of gear.

(I want to change the save system with a .json instead of txt)

Also implemented a function that let you choose if you want to keep doing things in the city or go back to kill monsters after performing any action