2.1
Created a function to be able to use potions.

Whenever the character is at 30% hp or below a prompt will appear where you can choose to use a potion or not.

Fixed a bug where everytime you start the game, having a save or not it would start the game.

Changed main file name from Juego.py to main.py